+++ changelog.md
@@ -2,25 +2,36 @@
 
 ## V 1.5.0
 
+### V 1.5.1
+
+- Added a button builder.
+- Added some missed exports
+
+### V 1.5.0
+
 - The version rewrites the old command option system making you the user control the order of options.
 - The docs should have got or should get an update soon.
+
+## V 1.4.0
 
 ### V 1.4.1
 
 - Fixed some oversights like descriptions optional on command options
 
-## V 1.4.0
+### V 1.4.0
 
 - Description localizations used setNameLocalizations instead of setDescriptionLocalizations
 - Added Default Command Permisions
 - Added NSFW tag
 - Added Contexts like if the comand is guild or groups or dm
 
-## V 1.3.1
+## V 1.3.0
+
+### V 1.3.1
 
 - Fixed the issue where the program was stoped without any way of continuing
 
-## V 1.3.0
+### V 1.3.0
 
 - Moved the command Record inside the client with module Augumentation.
 - Fixed a lot of bugs.